Initial content dropped in. Hugo builds.
Note that the teaching.md and teaching_notes.md files are currently not
rendered at all. This is by design. In Hugo's language, these are resources in
the branch bundles representing course, unit, lesson, or module.
Once I create templates for these respective content types, the
templates will automatically include these sections when the build is
configured to include them. (This allows building the site for different
audiences.)